refactor xds settings and logs
